requirements.txt
Django==3.2.20
django-environ==0.8.1
django-cors-headers==3.12.0
django_extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
